+++ NEWS.md
@@ -10,6 +10,9 @@ New language features
   and associativity as other arrow-like operators ([#36666]).
 * Compilation and type inference can now be enabled or disabled at the module level
   using the experimental macro `Base.Experimental.@compiler_options` ([#37041]).
+* The library name passed to `ccall` or `@ccall` can now be an expression involving
+  global variables and function calls. The expression will be evaluated the first
+  time the `ccall` executes ([#36458]).
 
 Language changes
 ----------------